+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 - Added `childstate_getKeysPaged` RPC
 - Align static metadata with latest Substrate
 - Adjust error message for contracts V3 (extrinsic from runtime)
+- Correct module optionality in `api.derive.balances.account`
 
 
 ## 4.14.1 Jun 14, 2021